refactor(components): move BookmarkButton into a topic folder
Promote src/components/BookmarkButton.tsx to
src/components/bookmarks/BookmarkButton.tsx to match the existing
topic-folder convention (editor/, providers/, ui/).
